Remove undefined symbol
PTR is not supported by gcc since 2022-05-10 given the changelog:
https://github.com/gcc-mirror/gcc/blob/master/include/ChangeLog

These seems to be an old symbol used for compatibility between non
ANSI-C and ANSI-C codes. I'm not sure it is still relevant to make the
distinction; if the library was to be compiled with an old enough
compiler that would require the "untyped pointer" to be declared as
`char *`, there would be many more issues elsewhere in the code.