Use for-loop instead of while-loop when parsing
Using for-loop constructs ease the following of boundary conditions.
This commit also uses more broadly standard string functions to handle
more edge cases.

In src/allocators.cxx, although it has been reworked, the
allocator-parsing loop has not been changed to a for-loop due to the non
linear advancement of `defs` which is incremented in the loop body, and
not only at the end of the iteration.