Check that the library is properly loaded when intercepting dlopen
In the interception of dlopen, the library may not be fully loaded when
the interception happens. For example, the loading of extra allocator
may require to dlopen an external library.
In case of an dlopen happening mid-initialisation of flexmalloc, a
segmentation fault will occur as codelocation is still NULL, and so is
the fallback allocator.

By checking whether the library is loaded or not, we only postpone the
update of code location to its original call, later in the
initialisation process. The code locations are still updated otherwise.