Add debug allocator for testing purposes
The debug allocator relies on the posix allocator for working, but
records its own statistics.

The chosen allocator in the tests is now the debug one, and the tests
are run against the debug version of the library. There is now a wrapper
script to run the tests. Maybe, the main script could test whether the
environment variable are already set, so we can have a single script for
everything.

In the tests/ and src/ makefiles, the parameters have been split between
CXXFLAGS and CPPFLAGS.